Extracted common functionality from DefaultI18NResourceProvider to a new class AbstractI18NResourceProvider Corrected package name on JAttachDetachListner
