+++ src/main/java/com/jain/addon/i18N/component/I18NUI.java
@@ -17,7 +17,7 @@ package com.jain.addon.i18N.component;
 
 import java.util.Locale;
 
-import com.jain.addon.I18N.listners.JAttachDetachListner;
+import com.jain.addon.i18N.listners.JAttachDetachListner;
 import com.jain.addon.event.EventHandler;
 import com.jain.addon.i18N.I18NChangeListener;
 import com.jain.addon.i18N.I18NHelper;
