+++ src/main/java/com/jain/addon/i18N/listners/JAttachDetachListner.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.jain.addon.I18N.listners;
+package com.jain.addon.i18N.listners;
 
 import java.lang.reflect.Method;
 import java.util.ArrayList;
