+++ src/main/java/com/jain/addon/resource/DefaultI18NResourceProvider.java
@@ -17,13 +17,11 @@ package com.jain.addon.resource;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.text.MessageFormat;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Properties;
 
-import com.jain.addon.JNINamed;
 import com.jain.addon.resource.I18NProvider;
 
 /**
@@ -33,7 +31,7 @@ import com.jain.addon.resource.I18NProvider;
  * @version 1.0.0
  */
 @SuppressWarnings("serial")
-public class DefaultI18NResourceProvider implements I18NProvider {
+public class DefaultI18NResourceProvider extends AbstractI18NResourceProvider {
 	private Map<String, Properties> propsForLocale;
 	private static I18NProvider instance;
 
@@ -78,75 +76,12 @@ public class DefaultI18NResourceProvider implements I18NProvider {
 		return properties;
 	}
 
-	public String getText(Locale locale, String identifier, Object... params) {
-		return getText(locale, identifier, BLANK, params);
-	}
-
-	public String getTitle(Locale locale, String identifier, Object... params) {
-		String value = getText(locale, identifier, TITLE_KEY, params);
-		if(identifier.equals(value))
-			value = getText(locale, identifier, params); 
-		return value;
-	}
-
-	public String getMessage(Locale locale, String identifier, Object... params) {
-		String value = getText(locale, identifier, MESSAGE_KEY, params);
-		if(identifier.equals(value))
-			value = getText(locale, identifier, params); 
-		return value;
-	}
-
+	@Override
 	public String getProperty(Locale locale, String key, String defaultValue) {
 		Properties properties = getProperties(locale == null ? null : locale.getLanguage());
 		if(properties != null) {
 			return properties.getProperty(key, defaultValue);
 		}
 		return defaultValue;
-	}
-
-	public String getProperty(Locale locale, String key) {
-		return getProperty(locale, key, key);
-	}
-
-	public String getText(Locale locale, String identifier, String sufix, Object... params) {
-		Object[] parsedParams = parseParams(locale, identifier, params);
-		
-		String defaultValue = identifier;
-		
-		if(identifier.contains(KEY_SPLITTER))
-			identifier = identifier.substring(0, identifier.indexOf(KEY_SPLITTER));
-
-		String result = getProperty(locale, identifier + sufix, defaultValue);
-		return MessageFormat.format(result, parsedParams);
-	}
-
-	private Object[] parseParams(Locale locale, String identifier, Object... params) {
-		if(params == null || params.length == 0) {
-			if(identifier.contains(KEY_SPLITTER)) {
-				String [] identifiers = identifier.split(KEY_SPLITTER);
-
-				Object [] parsedParams = new String [identifiers.length - 1];
-
-				for (int i = 1; i < identifiers.length; i++) {
-					parsedParams [i - 1] = getProperty(locale, identifiers [i]);
-				}
-
-				return parsedParams;
-			}
-			return null;
-		}
-
-		Object [] parsedParams = new String [params.length];
-
-		for (int i = 0; i < params.length; i++) {
-			Object object = params [i];
-			if (object instanceof JNINamed)
-				parsedParams [i] = ((JNINamed) object).getDisplayName();
-			else if (object instanceof String)
-				parsedParams [i] = getProperty(locale, (String)object);
-			else 
-				parsedParams [i] = object;
-		}
-		return parsedParams;
 	}
 }
